Fish

+++ monolithic_esp/main/packet_def/packet_def.h
@@ -6,14 +6,14 @@
 /**
  * Packet Header
  * @version 3
- * @param packet_id Is a unique identifier per packet
+ * @param packet_id Is a squential series that is incremented from each node
  * @param hop_count Decrementing until zero, then drop
  * @param flags Currently not used
  */
 typedef struct __attribute__((packed)) {
     uint8_t network_id;
     uint8_t orig_node_id;
-    uint16_t packet_id;
+    uint8_t packet_id;
     uint8_t hop_count;
     uint8_t flags;
 } packet_header_t;
@@ -32,7 +32,7 @@ typedef struct __attribute__((packed)) {
  * @param precipitation Gives the amount since last reading/packet, scaled with 10
  * @param wind_speed In mili meters per second
  * @param solar_output Solar panel power production in mA scaled with 4
- * @param bat_voltage The diff from 3.7 V in 0.01 V increments
+ * @param bat_voltage The diff from 3.5 V in 0.04 V increments
  */
 typedef struct __attribute__((packed)) {
     int32_t longitude;
@@ -40,12 +40,12 @@ typedef struct __attribute__((packed)) {
     uint8_t air_humidity;
     int16_t air_tempeture;     // Can be 0, Will just send a warning
     int16_t soil_tempeture[4]; // Can be 0, Will just send a warning
-    uint8_t soil_moisture;     // Can be 0
+    uint8_t soil_moisture;     // Can be 0, Assume out of range, Implemented
     int16_t pressure;          // Can be 0, Might work
     uint16_t spectrum;
     uint16_t precipitation; // Can be 0, how would i check? Should i base it wheter its been initated?
     uint16_t wind_speed;    // Can be 0, --//--
-    uint8_t solar_output;
+    uint8_t solar_output;   // Can be 0, Implemented
     int8_t bat_voltage;
 } sensor_payload_t;
